# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "EHPS API",
    "description" : "Equipment Health & Performance System — semiconductor fab equipment maintenance. Covers fab-coordinator machine/technician management, routine checkups, alert escalation and approval, repair assignment/completion, and the equipment health dashboard. All endpoints (except /api/auth/**) require a bearer JWT obtained from POST /api/auth/login, and are scoped by role: FAB_COORDINATOR, MANAGER, or TECHNICIAN. See docs/API-CONTRACT.md for the full endpoint contract.",
    "version" : "2.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/technician/repairs/alerts/{alertId}/completion" : {
      "post" : {
        "tags" : [ "Technician — Repairs" ],
        "summary" : "Complete a repair",
        "description" : "Marks the repair for the given alert as complete, recording the repair outcome. Requires role TECHNICIAN.",
        "operationId" : "completeRepair",
        "parameters" : [ {
          "name" : "alertId",
          "in" : "path",
          "description" : "Alert ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CompleteRepairRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCompleteRepairResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/fab/technicians" : {
      "get" : {
        "tags" : [ "Fab — Technicians" ],
        "summary" : "List all technicians",
        "description" : "Returns all technicians managed by the calling fab coordinator. Requires role FAB_COORDINATOR.",
        "operationId" : "getAllTechnicians",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTechnicianResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Fab — Technicians" ],
        "summary" : "Register a new technician",
        "description" : "Creates a new technician account managed by the calling fab coordinator. Requires role FAB_COORDINATOR.",
        "operationId" : "addTechnician",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddTechnicianRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTechnicianResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/technician/repairs/machines/{machineId}" : {
      "get" : {
        "tags" : [ "Technician — Repairs" ],
        "summary" : "Get machine details for a repair",
        "description" : "Returns machine details for a machine assigned to the calling technician for repair. Requires role TECHNICIAN.",
        "operationId" : "getMachineDetails",
        "parameters" : [ {
          "name" : "machineId",
          "in" : "path",
          "description" : "Machine ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMachineDetailsResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/technician/repairs/alerts" : {
      "get" : {
        "tags" : [ "Technician — Repairs" ],
        "summary" : "List my approved repair alerts",
        "description" : "Returns approved alerts assigned to the calling technician for repair. Requires role TECHNICIAN.",
        "operationId" : "getApprovedRepairAlerts",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListApprovedRepairAlertResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/technician/repairs/alerts/{alertId}" : {
      "get" : {
        "tags" : [ "Technician — Repairs" ],
        "summary" : "Get a single approved repair alert",
        "description" : "Returns a single approved repair alert assigned to the calling technician. Requires role TECHNICIAN.",
        "operationId" : "getApprovedRepairAlert",
        "parameters" : [ {
          "name" : "alertId",
          "in" : "path",
          "description" : "Alert ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseApprovedRepairAlertResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/technician/checkups/machines/{machineId}" : {
      "get" : {
        "tags" : [ "Technician — Checkups" ],
        "summary" : "Get machine details for a checkup",
        "description" : "Returns machine details for a machine assigned to the calling technician for checkup. Requires role TECHNICIAN.",
        "operationId" : "getMachineDetails_1",
        "parameters" : [ {
          "name" : "machineId",
          "in" : "path",
          "description" : "Machine ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMachineDetailsResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/manager/technicians" : {
      "get" : {
        "tags" : [ "Manager — Machines & Technicians" ],
        "summary" : "List all technicians",
        "description" : "Returns every technician across the fab. Requires role MANAGER.",
        "operationId" : "getAllTechnicians_1",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTechnicianResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/manager/machines" : {
      "get" : {
        "tags" : [ "Manager — Machines & Technicians" ],
        "summary" : "List all machines",
        "description" : "Returns every machine across all fab coordinators. Requires role MANAGER.",
        "operationId" : "getAllMachines",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListMachineResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/machine-types" : {
      "get" : {
        "tags" : [ "Machine Types" ],
        "summary" : "List all machine types",
        "description" : "Returns all machine types. Available to any authenticated role.",
        "operationId" : "getAllTypes",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListMachineTypeResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/machine-types/{typeId}/technicians" : {
      "get" : {
        "tags" : [ "Machine Types" ],
        "summary" : "List technicians qualified for a machine type",
        "description" : "Returns technicians qualified to service the given machine type. Requires role FAB_COORDINATOR.",
        "operationId" : "getTechniciansByType",
        "parameters" : [ {
          "name" : "typeId",
          "in" : "path",
          "description" : "Machine type ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTechnicianResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/machine-types/{typeId}/parameters" : {
      "get" : {
        "tags" : [ "Machine Types" ],
        "summary" : "Get parameter rules for a machine type",
        "description" : "Returns the checkup parameter rules configured for the given machine type. Available to any authenticated role.",
        "operationId" : "getParameters",
        "parameters" : [ {
          "name" : "typeId",
          "in" : "path",
          "description" : "Machine type ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListParameterRuleResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/machine-types/{typeId}/machines" : {
      "get" : {
        "tags" : [ "Machine Types" ],
        "summary" : "List machines of a given type",
        "description" : "Returns machines of the given type. Requires role FAB_COORDINATOR or MANAGER.",
        "operationId" : "getMachinesByType",
        "parameters" : [ {
          "name" : "typeId",
          "in" : "path",
          "description" : "Machine type ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListMachineResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiResponseTechnicianResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/TechnicianResponse"
          }
        }
      },
      "TechnicianResponse" : {
        "type" : "object",
        "properties" : {
          "technicianId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "speciality" : {
            "type" : "string"
          }
        }
      },
...
      "MachineResponse" : {
        "type" : "object",
        "properties" : {
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineCode" : {
            "type" : "string"
          },
          "typeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "typeName" : {
            "type" : "string"
          },
          "installDate" : {
            "type" : "string",
            "format" : "date"
          },
          "fabCoordinatorId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "CompleteRepairRequest" : {
        "type" : "object",
        "properties" : {
          "changesDone" : {
            "type" : "string"
          },
          "observations" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseCompleteRepairResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/CompleteRepairResponse"
          }
        }
      },
      "CompleteRepairResponse" : {
        "type" : "object",
        "properties" : {
          "alertId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineCode" : {
            "type" : "string"
          },
          "alertStatus" : {
            "type" : "string"
          },
          "repairRecorded" : {
            "type" : "boolean"
          },
          "historyRecorded" : {
            "type" : "boolean"
          }
        }
      },
...
      "AddTechnicianRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "email" : {
            "type" : "string",
            "pattern" : "^[A-Za-z0-9._%+-]+@ehps\\.com$"
          },
          "phone" : {
            "type" : "string",
            "pattern" : "^\\d{10}$"
          },
          "password" : {
            "type" : "string",
            "pattern" : "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$"
          },
          "speciality" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "name", "speciality" ]
      },
...
      "ApiResponseMachineDetailsResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/MachineDetailsResponse"
          }
        }
      },
      "MachineDetailsResponse" : {
        "type" : "object",
        "properties" : {
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineCode" : {
            "type" : "string"
          },
          "typeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "typeName" : {
            "type" : "string"
          },
          "installDate" : {
            "type" : "string",
            "format" : "date"
          },
          "fabCoordinatorId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ApiResponseListApprovedRepairAlertResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ApprovedRepairAlertResponse"
            }
          }
        }
      },
      "ApprovedRepairAlertResponse" : {
        "type" : "object",
        "properties" : {
          "alertId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineCode" : {
            "type" : "string"
          },
          "problemMeasure" : {
            "type" : "string"
          },
          "severity" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "raisedOn" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "ApiResponseApprovedRepairAlertResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/ApprovedRepairAlertResponse"
          }
        }
      },
...
      "ApiResponseListTechnicianResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TechnicianResponse"
            }
          }
        }
      },
      "ApiResponseListMachineResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MachineResponse"
            }
          }
        }
      },
...
      "ApiResponseListMachineTypeResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MachineTypeResponse"
            }
          }
        }
      },
      "MachineTypeResponse" : {
        "type" : "object",
        "properties" : {
          "typeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "typeName" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseListParameterRuleResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ParameterRuleResponse"
            }
          }
        }
      },
      "ParameterRuleResponse" : {
        "type" : "object",
        "properties" : {
          "paramIndex" : {
            "type" : "integer",
            "format" : "int32"
          },
          "paramName" : {
            "type" : "string"
          },
          "unit" : {
            "type" : "string"
          },
          "good" : {
            "type" : "string"
          },
          "warning" : {
            "type" : "string"
          },
          "bad" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}